move realtime flag into the common timer
This is one patch in a series of several changes. The batch
processing mode never been designed to work properly with multiple input midi
files.

In a first step we move the realtime flag into the common timer which also stores the batch time, now.
later changes will implement a guess of timing parameters and the actual batch processing.