Use generic parameter handling for retunings.
On the one hand this simplfies the logic as it removes duplicated code.
On the other hand we have to deal with two levels of nested argument
scopes for each retuning call with external parameters.

The main benefit is to provide a flexible infrastructure that can be reused
in order to handle other problems, e.g. parameters for tones, intervals, integers.