Do not create box menu entries for boxes without a logic window.
The box menu is updated as soon as the logic window gets opened.
So there is no need to add an entry for this box. On the other hand this
fixes updating of the box menu when a logic window is closed.
This is especially a bug when a box is closed.

Currently we do not reuse command ids, but this should not be necessary.

The current implementation has two memory leaks and an id leak.
All are documented in MutFrame::UpdateBoxMenu().