Reuse command ids in the box selection menu.
The unused command ids are saved in the same
map as the used command ids. This map can only grow and is
copied once per call to UpdateBoxMenu. This was the easiest
implementation it could be optimized but UpdateBoxMenu is usually
called during the response to some user interaction so it is not
too time critical.