Implement a debug version of intrusive_ptr on Linux.
This version prints stores a backtrace from the constructor and immediately
prints the return address of its constructor. During destruction it can
print a complete stack trace of the constructor when it is deleted after
the main cleanup has finished. Thus, certain memory leaks cen be debuged
detected much easier.