First steps for MIDI system common, MIDI meta, and event based routing.
All changes use the new event classes to process data.
As handling the meta, SysEx and system messages depends on whether
the device is a realtime device or saves the stream, handling of some
messages is different between the two device types. Other messages depend
on the MIDI file which might get mixed in the output – a reason for confusion.