Make the test-suite compatible to ThreadSanitizer
The wxWidgets library has some possible deadlock, which might never occur.
Neverthesless it prevents the test-suite from succeeding.
The patch implements a general workaround that allows both the testsuite to
finish as well as to avoid too generic suppression rules.

This does not mean that this goal is achieved by this patch, but the other
changes that will be checked in recently will construct a compilable state.